[BOOKINGSG-4231]BS][Part 3] Page paginations Page size dropdown button
