+++ src/pagination/types.ts
@@ -9,3 +9,11 @@ export interface PaginationsProps {
     showPageSizeChanger?: boolean | undefined;
     onPageChange?: ((page: number) => void) | undefined;
 }
+
+export interface DropdownProps {
+    id?: string | undefined;
+    "data-testid"?: string | undefined;
+    className?: string | undefined;
+    pageSize?: number | undefined;
+    onPageSizeChange?: ((page: number) => void) | undefined;
+}
